# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/resource-manage/v1/pictures": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图片"
                ],
                "summary": "创建图片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "待上传图片",
                        "name": "picture",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.PictureCreateReply"
                        }
                    }
                }
            }
        },
        "/resource-manage/v1/videos": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "创建视频",
                "parameters": [
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.VideoCreateReply"
                        }
                    }
                }
            }
        },
        "/resource-manage/v1/videos/multipart": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "视频"
                ],
                "summary": "批量创建视频",
                "parameters": [
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video1",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video2",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video3",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video4",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video5",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video6",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video7",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "待上传视频",
                        "name": "video8",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.VideoCreateReply"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "protocol.PictureCreateReply": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "protocol.VideoCreateReply": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}